at type precision on channels

+++ TFE_package/src/channels/nociceptor/K_AHP.jl
@@ -1,19 +1,19 @@
 
 # --------------------------- K_AHP --------------------------- #
 
-function zAHP_inf(V)
+function zAHP_inf(V::Float64)
     beta_z_AHP = 5
     gamma_z =  4
     return 1 / ( 1 + exp( (beta_z_AHP - V)/gamma_z ) ) 
 end
 
-function tau_zAHP(V)
+function tau_zAHP(V::Float64)
     return 100
 end
 
 # ---- #
 
-function dot_zAHP(V, zAHP)
+function dot_zAHP(V::Float64, zAHP::Float64)
     return (zAHP_inf(V)-zAHP)/tau_zAHP(V)
 end
